+++ client/src/Pages/MyPage/ContentArea/Contents/OrderList.tsx
@@ -1,0 +1,45 @@
+import React from "react";
+import styled from "styled-components";
+
+import Section from "../../Section";
+import Table from "../../Table";
+import RecentOrderRow from "../../Table/rows/RecentOrderRow";
+import { EmptyRow } from "../../Table/rows/EmptyRow";
+import { recent } from "@/shared/dummy";
+
+const OrderList = () => {
+  return (
+    <Wrapper>
+      <Section
+        title="주문목록/배송조회"
+        descrition="최근 30일 내에 주문한 주문정보입니다."
+        lineType="long1"
+      >
+        <Table
+          ths={[
+            "날짜/주문번호",
+            "상품명/옵션	",
+            "상품금액/수량",
+            "주문상태",
+            "확인리뷰",
+          ]}
+          ratio={[1, 3, 1, 0.5, 0.5]}
+        >
+          {recent.length !== 0 ? (
+            <EmptyRow colSpan={5} message="조회내역이 없습니다." />
+          ) : (
+            recent.map((re) => <RecentOrderRow {...re} />)
+          )}
+        </Table>
+      </Section>
+    </Wrapper>
+  );
+};
+
+const DeliveryTracker = () => {
+  return <Wrapper></Wrapper>;
+};
+
+const Wrapper = styled.div``;
+
+export default OrderList;
