+++ client/src/Pages/MyPage/ContentArea/Contents/index.ts
@@ -1,0 +1,17 @@
+import Root from "./Root";
+import OrderList from "./OrderList";
+import UserInfo from "./UserInfo";
+import WishList from "./WishList";
+import Question from "./Question";
+import Review from "./Review";
+
+const Contents = {
+  "/": Root,
+  "orderlist": OrderList,
+  "wishlist": WishList,
+  "userinfo": UserInfo,
+  "question": Question,
+  "review": Review,
+};
+
+export default Contents;
