+++ client/src/Pages/MyPage/ContentArea/index.tsx
@@ -1,12 +1,11 @@
 import styled from "styled-components";
-
 import Contents from "./Contents";
 
-const Content = ({ current }) => {
-  const C = Contents[current];
+const ContentArea = ({ current }) => {
+  const Content = Contents[current];
   return (
     <ContentWrapper>
-      <C />
+      <Content />
     </ContentWrapper>
   );
 };
@@ -15,4 +14,4 @@ const ContentWrapper = styled.div`
   flex: 1;
 `;
 
-export default Content;
+export default ContentArea;
